+++ src/network/index.js
@@ -6,3 +6,4 @@ export { ReactionNetwork } from './ReactionNetwork.js';
 export { ScaffoldNode } from './ScaffoldNode.js';
 export { ScaffoldReactionNode } from './ScaffoldReactionNode.js';
 export { ScaffoldNetwork } from './ScaffoldNetwork.js';
+export { MetabolicNetwork } from './MetabolicNetwork.js';